Fix wrong script_path under MINGW environment
Under MINGW environment, the vim script path is passed as "\c\path\to\directory"
when it is passed to python in autoload\youcompleteme.vim it will be interpreted
as %PWD%\c\path\to\directory instead of the correct path.

This patch fixes it.

Tested with the vim shipped with 'Git bash' from Git application.